Move code to CPG core and replace existing functions
