+++ cpg-analysis/src/main/kotlin/de/fraunhofer/aisec/cpg/query/FlowQueries.kt
@@ -28,117 +28,9 @@ package de.fraunhofer.aisec.cpg.query
 import de.fraunhofer.aisec.cpg.graph.*
 import de.fraunhofer.aisec.cpg.graph.AccessValues
 import de.fraunhofer.aisec.cpg.graph.Node
-import de.fraunhofer.aisec.cpg.graph.edges.Edge
-import de.fraunhofer.aisec.cpg.graph.edges.flows.CallingContextIn
-import de.fraunhofer.aisec.cpg.graph.edges.flows.CallingContextOut
-import de.fraunhofer.aisec.cpg.graph.edges.flows.ContextSensitiveDataflow
-import de.fraunhofer.aisec.cpg.graph.edges.flows.Dataflow
-import de.fraunhofer.aisec.cpg.graph.edges.flows.IndexedDataflowGranularity
-import de.fraunhofer.aisec.cpg.graph.statements.expressions.InitializerListExpression
 import de.fraunhofer.aisec.cpg.graph.statements.expressions.Literal
 import de.fraunhofer.aisec.cpg.graph.statements.expressions.Reference
 import kotlin.collections.all
-
-interface StepSelector {
-    /**
-     * Returns `true` if the given [edge] will be followed giving the provided [ctx] under the
-     * provided configuration.
-     */
-    fun followEdge(
-        currentNode: Node,
-        edge: Edge<Node>,
-        ctx: Context,
-        analysisDirection: AnalysisDirection? = null,
-    ): Boolean
-}
-
-/**
- * Determines how far we want to follow edges within the analysis. [maxSteps] defines the total
- * number of steps we will follow in the graph (unlimited depth if `null`).
- */
-sealed class AnalysisScope(val maxSteps: Int? = null) : StepSelector {
-    override fun followEdge(
-        currentNode: Node,
-        edge: Edge<Node>,
-        ctx: Context,
-        analysisDirection: AnalysisDirection?,
-    ): Boolean {
-        // Follow the edge if maxSteps is null or if maxSteps < ctx.steps
-        return this.maxSteps == null || maxSteps <= ctx.steps
-    }
-}
-
-/**
- * Only intraprocedural analysis. [maxSteps] defines the total number of steps we will follow in the
- * graph (unlimited depth if `null`).
- */
-class Intraprocedural(maxSteps: Int? = null) : AnalysisScope(maxSteps) {
-    override fun followEdge(
-        currentNode: Node,
-        edge: Edge<Node>,
-        ctx: Context,
-        analysisDirection: AnalysisDirection?,
-    ): Boolean {
-        // Follow the edge if we're still in the maxSteps range and not an edge across function
-        // boundaries.
-        return (this.maxSteps == null || ctx.steps < maxSteps) && edge !is ContextSensitiveDataflow
-    }
-}
-
-/**
- * Enable interprocedural analysis. [maxCallDepth] defines how many function calls we follow at most
- * (unlimited depth if `null`). [maxSteps] defines the total number of steps we will follow in the
- * graph (unlimited depth if `null`).
- */
-class Interprocedural(val maxCallDepth: Int? = null, maxSteps: Int? = null) :
-    AnalysisScope(maxSteps) {
-    override fun followEdge(
-        currentNode: Node,
-        edge: Edge<Node>,
-        ctx: Context,
-        analysisDirection: AnalysisDirection?,
-    ): Boolean {
-        // Follow the edge if we're still in the maxSteps range and (if maxCallDepth is null or the
-        // call stack is not deeper yet)
-        return (this.maxSteps == null || ctx.steps < maxSteps) &&
-            (maxCallDepth == null || ctx.callStack.depth < maxCallDepth)
-    }
-}
-
-/** Determines in which direction we follow the edges. */
-sealed class AnalysisDirection : StepSelector {
-    override fun followEdge(
-        currentNode: Node,
-        edge: Edge<Node>,
-        ctx: Context,
-        analysisDirection: AnalysisDirection?,
-    ): Boolean {
-        return true
-    }
-
-    abstract fun unwrapNextStepFromEdge(edge: Edge<Node>): Node
-}
-
-/** Follow the order of the EOG */
-class Forward() : AnalysisDirection() {
-    override fun unwrapNextStepFromEdge(edge: Edge<Node>): Node {
-        return edge.end
-    }
-}
-
-/** Against the order of the EOG */
-class Backward() : AnalysisDirection() {
-    override fun unwrapNextStepFromEdge(edge: Edge<Node>): Node {
-        return edge.start
-    }
-}
-
-/** In and against the order of the EOG */
-class Bidirectional() : AnalysisDirection() {
-    override fun unwrapNextStepFromEdge(edge: Edge<Node>): Node {
-        TODO("Not yet implemented")
-    }
-}
 
 /** Determines if the predicate must or may hold */
 enum class AnalysisType {
@@ -152,161 +44,6 @@ enum class AnalysisType {
      * property/requirement.
      */
     MAY,
-}
-
-/** Configures the sensitivity of the analysis. */
-sealed class AnalysisSensitivity : StepSelector {
-    operator fun plus(other: AnalysisSensitivity) = arrayOf(this, other)
-
-    operator fun List<AnalysisSensitivity>.plus(
-        other: AnalysisSensitivity
-    ): Array<AnalysisSensitivity> {
-        return arrayOf(*this.toTypedArray(), other)
-    }
-
-    operator fun plus(other: Array<AnalysisSensitivity>): Array<AnalysisSensitivity> {
-        return arrayOf(*other, this)
-    }
-}
-
-/** Consider the calling context when following paths (e.g. based on a call stack). */
-class ContextSensitive() : AnalysisSensitivity() {
-    override fun followEdge(
-        currentNode: Node,
-        edge: Edge<Node>,
-        ctx: Context,
-        analysisDirection: AnalysisDirection?,
-    ): Boolean {
-        return if (edge is ContextSensitiveDataflow && analysisDirection is Forward) {
-            // Forward analysis
-            if (edge.callingContext is CallingContextIn) {
-                // Push the call of our calling context to the stack
-                ctx.callStack.push((edge.callingContext as CallingContextIn).call)
-                true
-            } else if (ctx.callStack.isEmpty()) {
-                true
-            } else if (edge.callingContext is CallingContextOut) {
-                // We are only interested in outgoing edges from our current
-                // "call-in", i.e., the call expression that is on the stack.
-                ctx.callStack.popIfOnTop((edge.callingContext as CallingContextOut).call)
-            } else {
-                true
-            }
-        } else if (edge is ContextSensitiveDataflow && analysisDirection is Backward) {
-            // Backward analysis
-            if (edge.callingContext is CallingContextOut) {
-                // Push the call of our calling context to the stack
-                ctx.callStack.push((edge.callingContext as CallingContextOut).call)
-                true
-            } else if (ctx.callStack.isEmpty()) {
-                true
-            } else if (edge.callingContext is CallingContextIn) {
-                // We are only interested in outgoing edges from our current
-                // "call-in", i.e., the call expression that is on the stack.
-                ctx.callStack.popIfOnTop((edge.callingContext as CallingContextIn).call)
-            } else {
-                true
-            }
-        } else {
-            true
-        }
-    }
-}
-
-/**
- * Differentiate between fields, attributes, known keys or known indices of objects. This does not
- * include computing possible indices or keys if they are not given as a literal.
- */
-class FieldSensitive() : AnalysisSensitivity() {
-    override fun followEdge(
-        currentNode: Node,
-        edge: Edge<Node>,
-        ctx: Context,
-        analysisDirection: AnalysisDirection?,
-    ): Boolean {
-        return if (edge is Dataflow) {
-            if (
-                currentNode is InitializerListExpression &&
-                    analysisDirection?.unwrapNextStepFromEdge(edge) in currentNode.initializers &&
-                    edge.granularity is IndexedDataflowGranularity
-            ) {
-                // currentNode is the ILE, it is the child and the next (e.g. read
-                // from).
-                // We try to pop from the stack and only select the elements with the
-                // matching index.
-                ctx.indexStack.popIfOnTop(edge.granularity as IndexedDataflowGranularity)
-            } else if (
-                analysisDirection?.unwrapNextStepFromEdge(edge) is InitializerListExpression &&
-                    edge.granularity is IndexedDataflowGranularity
-            ) {
-                // CurrentNode is the child and nextDFG goes to ILE => currentNode's written
-                // to. Push to stack
-                ctx.indexStack.push(edge.granularity as IndexedDataflowGranularity)
-                true
-            } else {
-                true
-            }
-        } else {
-            true
-        }
-    }
-}
-
-/**
- * Also consider implicit flows during the dataflow analysis. E.g. if a condition depends on the
- * value we're interested in, different behaviors in the branches can leak data and thus, the
- * dependencies of this should also be flagged.
- */
-class Implicit() : AnalysisSensitivity() {
-    override fun followEdge(
-        currentNode: Node,
-        edge: Edge<Node>,
-        ctx: Context,
-        analysisDirection: AnalysisDirection?,
-    ): Boolean {
-        return true
-    }
-}
-
-/**
- * Returns an instance of [FulfilledAndFailedPaths] where [FulfilledAndFailedPaths.fulfilled]
- * contains all possible shortest data flow paths (with [FullDataflowGranularity]) between the
- * starting node [this] and the end node fulfilling [predicate]. The paths are represented as lists
- * of nodes. Paths which do not end at such a node are included in [FulfilledAndFailedPaths.failed].
- *
- * Hence, if "fulfilled" is a non-empty list, a data flow from [this] to such a node is **possible
- * but not mandatory**. If the list "failed" is empty, the data flow is mandatory.
- */
-fun Node.followNextDFGEdgesUntilHitNew(
-    collectFailedPaths: Boolean = true,
-    findAllPossiblePaths: Boolean = true,
-    direction: AnalysisDirection = Forward(),
-    vararg sensitivities: AnalysisSensitivity = FieldSensitive() + ContextSensitive(),
-    scope: AnalysisScope = Interprocedural(),
-    earlyTermination: (Node, Context) -> Boolean = { _, _ -> false },
-    predicate: (Node) -> Boolean,
-): FulfilledAndFailedPaths {
-
-    return this.followXUntilHit(
-        x = { currentNode, ctx, path ->
-            if (direction is Forward) {
-                    currentNode.nextDFGEdges
-                } else {
-                    currentNode.prevDFGEdges
-                }
-                .filter { edge ->
-                    // TODO: There's a problem with the context which should be copied for each
-                    // (potential) path.
-                    scope.followEdge(currentNode, edge, ctx) &&
-                        sensitivities.all { it.followEdge(currentNode, edge, ctx) }
-                }
-                .map { direction.unwrapNextStepFromEdge(it) }
-        },
-        collectFailedPaths = collectFailedPaths,
-        findAllPossiblePaths = findAllPossiblePaths,
-        earlyTermination = earlyTermination,
-        predicate = predicate,
-    )
 }
 
 /**
@@ -345,7 +82,7 @@ fun dataFlow(
             }
             .fold(FulfilledAndFailedPaths(listOf(), listOf())) { result, direction ->
                 result +
-                    startNode.followNextDFGEdgesUntilHitNew(
+                    startNode.followDFGEdgesUntilHit(
                         collectFailedPaths = collectFailedPaths,
                         findAllPossiblePaths = findAllPossiblePaths,
                         direction = direction,
